+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Security in case of vulnerabilities.
 
 ## Unreleased  
 ### Added
+ - New route `/projects/all/by-month/conversions/{month}/{year}`
+ - New route `/projects/all/by-month/transfers/{month}/{year}`
+ - New route `/projects/all/by-month/conversions/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
+ - New route `/projects/all/by-month/transfers/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
  - New route `/projects/team/completed`
  - New route `/projects/team/new`
  - New route `/projects/team/handed-over`
@@ -27,13 +31,14 @@ Security in case of vulnerabilities.
  - Your team projects "By user" > "User" list (`/projects/team/users/{userId}`)
 
 ### Changed
-- Add an "orderBy" argument to the `ListAllProjectsByFilter` query
-- Allow `ListAllProjectsByFilter` query to handle multiple filters
+ - Add an "orderBy" argument to the `ListAllProjectsByFilter` query
+ - Allow `ListAllProjectsByFilter` query to handle multiple filters
+ - Filter out any local authorities with no projects in `ListAllProjectByLocalAuthorities`
 
 ### Fixed
  - Project for user list should show month and year (not day)
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/main...production-2025-04-17.164) for everything awaiting release
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-04-17.164...main) for everything awaiting release
 
 ---